Support plugin loading from bootstrap environment
