+++ core/src/test/java/org/teavm/vm/PluginLoaderTest.java
@@ -56,7 +56,7 @@ public class PluginLoaderTest {
     }
 
     private List<String> order(Class<?>... classes) {
-        return TeaVMPluginLoader.orderPlugins(PluginLoaderTest.class.getClassLoader(),
+        return TeaVMPluginReader.orderPlugins(PluginLoaderTest.class.getClassLoader(),
                 Arrays.stream(classes).map(Class::getName).collect(Collectors.toSet()));
     }
